+++ app/lib/utils/logger.js
@@ -1,11 +1,2 @@
-function info(message) {
-  // eslint-disable-next-line no-console
-  console.log(message);
-}
-
-function error(message) {
-  // eslint-disable-next-line no-console
-  console.error(message);
-}
-
-module.exports = { info, error };
+const config = require('../config');
+module.exports = require('nhsuk-bunyan-logger')(config.app.name);